refactor(bpa-server): move BPSec key loading out of config/ into bpsec
  Address review: config/ should hold only configuration, mirroring the
  config/static_routes split. Move the JWKS loading, key/binding validation,
  and key-file permission check (with their tests) out of config/bpsec.rs into
  bpsec.rs as PatternKeySource::from_config; config/bpsec.rs is now just the
  serde Config and KeyBindingConfig structs. main.rs and watch_keys call
  PatternKeySource::from_config directly. No behaviour change.

  Document why KeyProvider and KeySource are two traits: KeyProvider selects the
  key material for a bundle flow (and may inspect the bundle source or a
  security-context block), while KeySource resolves an individual key from the
  security source EID, which under RFC 9172 need not be the bundle source. Note
  on CurrentKeys explains why the snapshot shares the table via Arc (whole-table
  provider on the per-bundle parse path) rather than cloning.
